Refactor ExplanationCard into its own file
