+++ src/components/calculator/ExplanationCard/riskTolerancePopover.tsx
@@ -1,0 +1,55 @@
+import React from 'react'
+import { Popover } from 'react-bootstrap'
+import { Link } from 'react-router-dom'
+
+export const riskTolerancePopover = (
+  <Popover id="popover-basic">
+    <Popover.Title as="h3">About At Risk Populations</Popover.Title>
+    <Popover.Content>
+      <p>
+        Our living group (all at average risk) has agreed to a{' '}
+        <Link to="/paper/2-riskiness">1% risk of getting COVID per year. </Link>
+      </p>
+      <p>
+        We suggest more caution (0.1% risk per year) for people who are more
+        vulnerable to severe illness from COVID (or in contact with more
+        vulnerable people)
+      </p>
+      <p>
+        <p>
+          <a
+            href="https://www.cdc.gov/coronavirus/2019-ncov/need-extra-precautions/older-adults.html"
+            target="_blank"
+            rel="noreferrer"
+          >
+            Vulnerability increases with age.
+          </a>{' '}
+          We think age over 60 confers substantial increased vulnerability to
+          severe illness from COVID.
+        </p>
+        <p>
+          <a
+            href="https://www.cdc.gov/coronavirus/2019-ncov/need-extra-precautions/people-with-medical-conditions.html"
+            target="_blank"
+            rel="noreferrer"
+          >
+            Certain underlying medical conditions
+          </a>{' '}
+          also confer increased vulnerability:
+        </p>
+        <ul>
+          <li>BMI of 30 or higher</li>
+          <li>Type 2 diabetes mellitus</li>
+          <li>Smoking</li>
+          <li>COPD or other heart conditions</li>
+          <li>Cancer</li>
+          <li>Chronic kidney disease</li>
+          <li>Immunocompromise from solid organ transplant</li>
+          <li>Sickle cell disease</li>
+        </ul>
+      </p>
+    </Popover.Content>
+  </Popover>
+)
+
+export default riskTolerancePopover
